Moved Dimension, PartialShape, Interval, Rank to ov namespace
