+++ ngraph/core/include/ngraph/rank.hpp
@@ -5,11 +5,8 @@
 #pragma once
 
 #include "ngraph/dimension.hpp"
+#include "openvino/core/rank.hpp"
 
 namespace ngraph {
-/// \brief Alias for Dimension, used when the value represents the number of axes in a shape,
-///        rather than the size of one dimension in a shape.
-///
-/// XXX: THIS TYPE IS EXPERIMENTAL AND THE ENTIRE DESIGN IS SUBJECT TO CHANGE.
-using Rank = Dimension;
+using ov::Rank;
 }  // namespace ngraph
